Refactor UI components and event handling: Update FXML files for consistent styling and layout adjustments. Introduce `ActionButton.css` for reusable button styles. Enhance mixer and gain event handling by adding application state context. Simplify `AudioStream` by removing unused comments and refining logic.

+++ src/main/java/net/fg83/hytalkclient/audio/AudioStream.java
@@ -6,59 +6,26 @@ import net.fg83.hytalkclient.util.AppConstants;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
-/**
- * Abstract base class for all audio streams in the application.
- *
- * Provides common functionality:
- * - Audio level tracking for metering
- * - Gain control
- * - Mute control
- * - Running state management
- */
 public abstract class AudioStream {
 
-    // Audio level monitoring (0.0 - 1.0)
     protected final AtomicReference<Float> currentLevel = new AtomicReference<>(0.0f);
 
-    // Gain control (0.0 - 1.25, where 1.0 = unity)
     protected final AtomicReference<Float> gain = new AtomicReference<>(1.0f);
 
-    // Proximity attenuation trim (0.0 - 1.0, where 1.0 = full volume, independent of fader gain)
     protected final AtomicReference<Float> attenuation = new AtomicReference<>(0.0f);
 
-    // Mute state
     protected final AtomicBoolean muted = new AtomicBoolean(false);
 
-    // Running state (for streams with lifecycle)
     protected final AtomicBoolean running = new AtomicBoolean(false);
 
-    /**
-     * Get the current audio level for metering.
-     * Thread-safe.
-     *
-     * @return Current audio level (0.0 - 1.0)
-     */
     public float getCurrentLevel() {
         return muted.get() ? 0.0f : currentLevel.get();
     }
 
-    /**
-     * Set the gain multiplier.
-     *
-     * @param gain Gain multiplier (0.0 - 1.25, where 1.0 = unity)
-     */
     public void setGain(float gain) {
         this.gain.set(Math.max(0.0f, Math.min(1.25f, gain)));
     }
-
-    /**
-     * Get the current gain multiplier.
-     *
-     * @return Current gain (0.0 - 1.25)
-     */
-    public float getGain() {
-        return gain.get();
-    }
+    public float getGain() { return gain.get(); }
 
     public void setAttenuation(float attenuation) {
         this.attenuation.set(Math.max(0.0f, Math.min(1.0f, attenuation)));
@@ -67,52 +34,23 @@ public abstract class AudioStream {
         return attenuation.get();
     }
 
-
-    /**
-     * Set the mute state.
-     *
-     * @param muted true to mute, false to unmute
-     */
     public void setMuted(boolean muted) {
         this.muted.set(muted);
     }
 
-    /**
-     * Get the mute state.
-     *
-     * @return true if muted, false otherwise
-     */
     public boolean isMuted() {
         return muted.get();
     }
 
-    /**
-     * Check if the stream is currently running/active.
-     *
-     * @return true if running, false otherwise
-     */
     public boolean isRunning() {
         return running.get();
     }
 
-    /**
-     * Update the current audio level with decay for smooth metering.
-     * Should be called by subclasses when processing audio.
-     *
-     * @param newLevel The new peak level from current frame (0.0 - 1.0)
-     */
     protected void updateLevel(float newLevel) {
         float decayed = currentLevel.get() * 0.95f;
         currentLevel.set(Math.max(newLevel, decayed));
     }
 
-    /**
-     * Calculate peak level from 16-bit PCM samples.
-     * Utility method for subclasses.
-     *
-     * @param samples PCM sample bytes (little-endian 16-bit)
-     * @return Peak level (0.0 - 1.0)
-     */
     protected float calculatePeakLevel(byte[] samples) {
         float maxLevel = 0.0f;
 
@@ -124,14 +62,6 @@ public abstract class AudioStream {
         return maxLevel;
     }
 
-    /**
-     * Apply gain to a single sample value and clamp to prevent clipping.
-     * Utility method for subclasses.
-     *
-     * @param sample Input sample value
-     * @param gain Gain multiplier
-     * @return Gained and clamped sample
-     */
     protected short applyGainToSample(short sample, float gain) {
         float gained = sample * gain * (isMuted() ? 0.0f : 1.0f);
         gained = Math.max(-32768, Math.min(32767, gained));
@@ -141,6 +71,6 @@ public abstract class AudioStream {
     protected float applyGainToSample(float sample, float gain) {
         float gained = sample * gain * attenuation.get() * (isMuted() ? 0.0f : 1.0f);
         gained = Math.max(-32768, Math.min(32767, gained));
-        return (float) gained;
+        return gained;
     }
 }
